Change the capture-replay logic in GaudiStableDiffusionPipeline
