+++ optimum/habana/diffusers/pipeline_utils.py
@@ -123,10 +123,8 @@ class GaudiDiffusionPipeline(DiffusionPipeline):
                     error.msg = f"Could not import habana_frameworks.torch. {error.msg}."
                     raise error
                 self.ht = ht
-                self.hpu_graph = ht.hpu.HPUGraph()
                 self.hpu_stream = ht.hpu.Stream()
-                self.static_inputs = list()
-                self.static_outputs = None
+                self.cache = {}
             else:
                 try:
                     import habana_frameworks.torch.core as htcore
